Modified architecture folders
