+++ public/controllers/agent/register-agent/services/index.ts
@@ -1,1 +1,3 @@
 export * from './packages-service'
+export * from './register-agent-service';
+export * from './steps-service';
