+++ public/controllers/agent/register-agent/steps/agent-name-groups/agent-group.tsx
