+++ public/controllers/agent/register-agent/steps/server-address/server-address-content.tsx
@@ -1,0 +1,23 @@
+import React, { Fragment } from 'react';
+import ServerAddress from './server-address';
+
+/**
+ * 
+ * @param title 
+ * @param state 
+ * @param onChange 
+ */
+export const getServerAddressStepContent = (
+    title: string,
+    state: any,
+    onChange: (field: string, value: string) => void,
+  ) => {
+    return (
+      <Fragment>
+        <ServerAddress
+          defaultValue={state.serverAddress || ''}
+          onChange={value => onChange('serverAddress', value)}
+        />
+      </Fragment>
+    );
+  };
