+++ public/controllers/agent/register-agent/steps/wazuh-password/wazuh-password-content.tsx
@@ -1,0 +1,17 @@
+import React, { Fragment } from "react";
+import WazuhPassword from "./wazuh-password";
+
+export const getWazuhPasswordStepContent = (
+    title: string,
+    state: any,
+    onChange: (field: string, value: string) => void,
+  ) => {
+    return (
+      <Fragment>
+        <WazuhPassword
+          defaultValue={state.wazuhPassword}
+          onChange={(value: string) => onChange('wazuhPassword', value)}
+        />
+      </Fragment>
+    );
+  };
