+++ public/controllers/agent/register-agent/steps/wazuh-password/wazuh-password.tsx
